Simplify calltree tests to use snapshots
